fix(rust, python): fix schema of decimal type reads
